+++ cli/src/dag/dag_validate.rs
@@ -53,7 +53,11 @@ pub(crate) async fn validate_dag(path: PathBuf) -> AnyResult<Dag, NexusCliError>
 
 #[cfg(test)]
 mod tests {
-    use {super::*, assert_matches::assert_matches, nexus_sdk::types::Dag};
+    use {
+        super::*,
+        assert_matches::assert_matches,
+        nexus_sdk::types::{Dag, FromPort},
+    };
 
     // == Various graph shapes ==
 
@@ -221,6 +225,33 @@ mod tests {
         let res = validate(dag);
 
         assert_matches!(res, Err(e) if e.to_string().contains("'Vertex: a' is not connected to the DAG."));
+    }
+
+    #[test]
+    fn test_encrypted_port_output_valid() {
+        let dag: Dag =
+            serde_json::from_str(include_str!("_dags/encrypted_port_output_valid.json")).unwrap();
+
+        assert!(dag.edges.first().unwrap().from.encrypted);
+        assert_eq!(
+            *dag.outputs.unwrap().first().unwrap(),
+            FromPort {
+                vertex: "b".to_string(),
+                output_variant: "1".to_string(),
+                output_port: "1.0".to_string(),
+                encrypted: true,
+            }
+        )
+    }
+
+    #[test]
+    fn test_output_and_edge_invalid() {
+        let dag: Dag =
+            serde_json::from_str(include_str!("_dags/output_and_edge_invalid.json")).unwrap();
+
+        let res = validate(dag);
+
+        assert_matches!(res, Err(e) if e.to_string().contains("'Vertex: a' cannot have both outgoing edges and ports marked as output."));
     }
 
     #[test]