+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tool registration now takes `invocation_cost` parameter and returns 2 owner caps `OverTool` and `OverGas`
 - `nexus conf --nexus.objects` is now the only way to populate the `nexus.objects` field in the config
 - `nexus conf` changed to have `set` and `get` subcommands
+- `nexus dag execute` now takes `--encrypt` argument that accepts `vertex.port` pairs to encrypt before sending data on-chain
+- JSON DAG now accepts `encrypted` field on `edges.[].from`
 
 ### `nexus-sdk`